added user registration page and password encryption

+++ admin/includes/view_all_posts.php
@@ -106,7 +106,8 @@ if (isset($_POST['checkBoxArray'])) {
         echo "<td>$post_date</td>";
         echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
         echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>EDIT</a></td>";
-        echo "<td><a href='posts.php?delete={$post_id}'>DELETE</a></td>";
+        echo "<td><a onclick=\"confirm('Are you sure you would like to delete this file')\" href='posts.php?delete={$post_id}'>DELETE</a></td>";
+
         echo "</tr>";
 
       }
